# app-backend/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OpenFabric Assignment API Documentation
  version: 1.0.0
paths:
  /products:
    get:
      summary: Get all products
      responses:
        '200':
          description: OK
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                name:
                  type: string
                quantity:
                  type: number
                price:
                  type: number
                model:
                  type: string
                color:
                  type: string
      responses:
        '201':
          description: Created
  /products/{id}:
    get:
      summary: Get a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '404':
          description: Product not found
    put:
      summary: Update a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: number
                name:
                  type: string
                quantity:
                  type: number
                price:
                  type: number
                model:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: OK
        '404':
          description: Product not found
    delete:
      summary: Delete a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content